ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
websockets==12.0
python-socketio==5.10.0
jinja2==3.1.2
requests==2.31.0
